fix/corrigido  no metodo de inserir produtos a uma venda aberta. Alterado tipagem de qtd para double

+++ src/main/java/com/hortifruti/hortifrutiapi/dto/estoque/EstoqueProdutoRequestDTO.java
@@ -5,5 +5,5 @@ import java.util.UUID;
 public record EstoqueProdutoRequestDTO(
         UUID produtoId,
         UUID sedeId,
-        Integer quantidade
+        Double quantidade
 ) {}
