+++ src/main/java/com/hortifruti/hortifrutiapi/model/EstoqueProduto.java
@@ -34,5 +34,5 @@ public class EstoqueProduto implements Serializable {
 
     @NotNull
     @Min(0)
-    private Integer quantidade;
+    private Double quantidade;
 }
